pyproject.toml
[tool.poetry]
name = "drf-todo-app"
version = "0.1.0"
description = ""
authors = ["Sohaib-50 <sohaibahmedabbasi2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
